+++ internal/hv/embed_darwin_amd64.go
@@ -1,8 +1,0 @@
-//go:build embed_hv && darwin && amd64
-
-package hv
-
-import _ "embed"
-
-//go:embed assets/proton-hv-darwin-amd64
-var helperBinary []byte