+++ internal/hv/embed_linux_arm64.go
@@ -1,8 +1,0 @@
-//go:build embed_hv && linux && arm64
-
-package hv
-
-import _ "embed"
-
-//go:embed assets/proton-hv-linux-arm64
-var helperBinary []byte